requirements.txt
# Python dependencies for this project
# Install with: pip install -r requirements.txt

ultralytics==8.3.196
torch==2.8.0
torchvision==0.23.0
opencv-python==4.12.0.88
numpy==2.2.6
pillow==11.3.0
PyYAML==6.0.2
requests==2.32.5
psutil==7.0.0
vietocr